NEW Move form schema logic into FormSchemaController
